fix CS0172

+++ OpenRA.Mods.Common/Traits/World/TerrainGeometryOverlay.cs
@@ -48,7 +48,7 @@ namespace OpenRA.Mods.Common.Traits
 				var height = (int)map.MapHeight.Value[uv];
 				var tile = map.MapTiles.Value[uv];
 				var ti = tileSet.GetTileInfo(tile);
-				var ramp = ti != null ? ti.RampType : 0;
+				var ramp = ti != null ? (int)ti.RampType : 0;
 
 				var corners = map.CellCorners[ramp];
 				var color = corners.Select(c => colors[height + c.Z / 512]).ToArray();
